% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{As numeric}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{character vector with numeric values where comma is used as thousand seperator}
}
\description{
As numeric
}
