% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_change}
\alias{f_change}
\title{This function is used to calculate price change in \%}
\usage{
f_change(x, digits = 5)
}
\arguments{
\item{x}{stock price}
}
\value{
A vector with prince changes in \%
}
\description{
This function is used to calculate price change in \%
}
