% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_symbol.R
\name{f_get_stock}
\alias{f_get_stock}
\title{This function returns company name code for a given ticker code}
\usage{
f_get_stock(ticker, index = NULL)
}
\arguments{
\item{ticker}{ticker code}

\item{index}{Index name. Supply either ticker code or index, not both. Index can be a vector.}
}
\value{
A vector with company names matching the input name(ticker or index).
}
\description{
This function returns company name code for a given ticker code
}
