% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_symbol.R
\name{f_get_symbol}
\alias{f_get_symbol}
\title{This function returns ticker code for a stock}
\usage{
f_get_symbol(stock = "NOVO", index = NULL, output = "vector")
}
\arguments{
\item{stock}{Stock name. Full stock name not needed as ticker codes for stock names like parameter input is returned.}

\item{index}{Index name. Supply either stock name or index, not both.}
}
\value{
A vector with ticker codes matching the input name(stock or index). If the output parameter is not "vector", then a data.frame with ticker codes and stock names matching input stock name is returned.
}
\description{
This function returns ticker code for a stock
}
