% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_indicator.R
\name{f_indicator}
\alias{f_indicator}
\title{This function calculates common indicators such as moving averages and RSI}
\usage{
f_indicator(data, n = 10, m = 5)
}
\arguments{
\item{n}{number of days used to calculate the slope of moving average. Default is 10}

\item{m}{number of days used to calculate moving average. Default is 5}

\item{ticker}{a data.table obtained from f_load}
}
\value{
A data.table
}
\description{
This function calculates common indicators such as moving averages and RSI
}
