% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_indicator.R
\name{f_lag}
\alias{f_lag}
\title{Lag function}
\usage{
f_lag(x, n = 1)
}
\arguments{
\item{x}{A vector to perform lag on}

\item{n}{number of elements to shift. Default is 1.}
}
\description{
Lag function
}
\examples{
f_lag(x = 1:10, n = 1)

}
