% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_load}
\alias{f_load}
\title{This function loads data from Yahoo.}
\usage{
f_load(ticker, from_date = "2014-01-01")
}
\arguments{
\item{ticker}{ticker code from Yahoo. Provide a vector of ticker codes to get multiple stocks.}

\item{from_date}{loads data from the date untill today. Date format is yyyy-mm-dd.}
}
\value{
A data.table
}
\description{
This function loads data from Yahoo.
}
