% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_load_one}
\alias{f_load_one}
\title{This function loads one stock from Yahoo.}
\usage{
f_load_one(ticker, from_date = "2014-01-01")
}
\arguments{
\item{ticker}{ticker code from Yahoo. Use f_load to multiple stocks.}

\item{from_date}{loads data from the date untill today. Date format is yyyy-mm-dd.}
}
\value{
A data.table
}
\description{
This function loads one stock from Yahoo.
}
