% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_indicator.R
\name{f_slope}
\alias{f_slope}
\title{This function calculates the slope of a moving average}
\usage{
f_slope(data)
}
\arguments{
\item{data}{data.table. data should be a obj an outcome from f_indicators to ensure MA is present in data.}
}
\value{
The slope of the moving average
}
\description{
This function calculates the slope of a moving average
}
