% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ticker.R
\name{f_ticker}
\alias{f_ticker}
\title{This function returns the ticker for a stock}
\usage{
f_ticker(stock)
}
\arguments{
\item{stock}{A string providing the stock name}
}
\value{
A string with the ticker code
}
\description{
This function returns the ticker for a stock
}
