% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_load.R
\name{f_try_catch}
\alias{f_try_catch}
\title{This function is an extension of tryCatch}
\usage{
f_try_catch(expr)
}
\arguments{
\item{expr}{expression to evalute}
}
\value{
A list with value, warning and error message
}
\description{
This function is an extension of tryCatch
}
