% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_symbol.R
\name{f_update_symbol}
\alias{f_update_symbol}
\title{This function update the lists of symbols available from a given exchange/in an index.}
\usage{
f_update_symbol(str_exchange_index = "NASDAQ")
}
\arguments{
\item{str_exhange_index}{A string specifying which exhange/index to look at. Possible values: NASDAQ, GDAXIOMX, OMX, OMXC25 and OMXCXC20GI.}
}
\value{
Nothing is returned. The function saves the symbols as .txt in ./data/symbols folder
}
\description{
This function update the lists of symbols available from a given exchange/in an index.
}
