% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sectors.R
\name{get_industries}
\alias{get_industries}
\title{Get industries}
\usage{
get_industries(country)
}
\arguments{
\item{country}{character string with country name (for instance "usa")}
}
\value{
character vector with industry names
}
\description{
Get industries
}
