% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sectors.R
\name{get_sectors}
\alias{get_sectors}
\title{Get sectors}
\usage{
get_sectors(country)
}
\arguments{
\item{country}{character string with country name (for instance "usa")}
}
\value{
character vector with sector names
}
\description{
Get sectors
}
