% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data from Yahoo.}
\usage{
load_data(ticker, from = "2014-01-01")
}
\arguments{
\item{ticker}{ticker code from Yahoo.}

\item{from}{load data `from` untill today where the format of `from` is yyyy-mm-dd.}
}
\value{
A data.frame
}
\description{
Load data from Yahoo.
}
\examples{
# Load NOVO data since 2021-01-01
data <- load_data(ticker = "NOVO-B.CO", from = "2021-01-01")
data
}
