% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tradingview_pe.R
\name{load_tradingview_pe}
\alias{load_tradingview_pe}
\title{Load tradingview pe}
\usage{
load_tradingview_pe(url)
}
\arguments{
\item{url}{url (for instance https://www.tradingview.com/markets/stocks-usa/sectorandindustry-sector/commercial-services/)}
}
\value{
numeric vector with PE for the given sector/industry
}
\description{
Load tradingview pe
}
