% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yahoo_eps.R
\name{load_yahoo_eps}
\alias{load_yahoo_eps}
\title{Load yahoo eps}
\usage{
load_yahoo_eps(ticker)
}
\arguments{
\item{ticker}{ticker code}
}
\value{
numeric value (one dimensional)
}
\description{
Load yahoo eps
}
