% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yahoo_pe.R
\name{load_yahoo_pe}
\alias{load_yahoo_pe}
\title{Load yahoo pe}
\usage{
load_yahoo_pe(ticker, verbose = TRUE)
}
\arguments{
\item{ticker}{character vector with ticker codes}

\item{verbose}{logical indicating if loop counter should be on (TRUE) or off (FALSE)}
}
\value{
numeric vector with PE
}
\description{
Load yahoo pe
}
