% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_change.R
\name{price_change}
\alias{price_change}
\title{Daily price change in percent}
\usage{
price_change(x, digits = 5)
}
\arguments{
\item{x}{stock price}

\item{digits}{number of digits used in rounding}
}
\value{
A numerical vector with prince changes in percent
}
\description{
Daily price change in percent
}
