% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{prince_change_from_start}
\alias{prince_change_from_start}
\title{Change from start}
\usage{
prince_change_from_start(x, digits = 4)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits using in rounding}
}
\description{
Change from start
}
