% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{prince_change_since_start}
\alias{prince_change_since_start}
\title{Change from start}
\usage{
prince_change_since_start(x, start = 1, digits = 4)
}
\arguments{
\item{x}{numeric vector}

\item{start}{integer specifying start element (default is 1, i.e. the first element of x)}

\item{digits}{number of digits using in rounding}
}
\value{
a numeric vector of same length as x
}
\description{
Change from start
}
