% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns in data.frame}
\usage{
rename_cols(data, ticker, rename = TRUE)
}
\arguments{
\item{data}{data.frame with columns Close, Change and ChangeSinceStart}

\item{ticker}{character string with ticker name}

\item{rename}{logical (default is TRUE). If FALSE, data is returned without any renaming}
}
\value{
data.frame
}
\description{
Renaming columns from:
* Close -> ticker
* Change -> ticker_Change
* ChangeSinceStart -> ticker_ChangeSinceStart
}
\details{
This is a helper function for `load_data`
}
