% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_yahoo_pe.R
\name{replace_comma_numerically}
\alias{replace_comma_numerically}
\title{Replace comma numerically}
\usage{
replace_comma_numerically(x)
}
\arguments{
\item{x}{vector of any class}
}
\value{
numeric value
}
\description{
Replace comma numerically
}
