% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data from Yahoo.}
\usage{
load_data(ticker, src = "yahoo", from = "2014-01-01", prefix = TRUE)
}
\arguments{
\item{ticker}{ticker code from Yahoo.}

\item{src}{source name ("yahoo" or "FRED")}

\item{from}{load data `from` untill today where the format of `from` is yyyy-mm-dd.}

\item{prefix}{logical (default is TRUE) indicating if ticker name should be included in column name of columns Change and ChangeSinceStart (only applicable for FRED data)}
}
\value{
A data.frame
}
\description{
Load data from Yahoo.
}
\examples{
# Load NOVO data since 2021-01-01
data <- load_data(ticker = "NOVO-B.CO", from = "2021-01-01")
data

# Load Market Yield on U.S. Treasury Securities at 10-Year Constant Maturity
data <- load_data(ticker = "DGS10", src = "FRED")
data
}
